# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Golug Demo Code",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/distros/list": {
            "get": {
                "description": "List all users",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/": {
            "post": {
                "description": "Create a New User",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                }
            }
        },
        "/api/v1/users/list": {
            "get": {
                "description": "List all users",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "description": "Retrieves a User by ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                }
            },
            "put": {
                "description": "Create a New User",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                }
            },
            "delete": {
                "description": "Retrieves a User by ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmodels.LinuxUser"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dbmodels.LinuxUser": {
            "type": "object",
            "properties": {
                "linuxDistro": {
                    "type": "string"
                },
                "linuxUserID": {
                    "type": "integer"
                },
                "linuxUserName": {
                    "type": "string"
                }
            }
        }
    }
}